requirements.txt
python-dotenv==1.1.1
pandas==2.3.2 
numpy==2.3.2
boto3==1.39.11
pyYAML==6.0.2
scikit-learn==1.7.1
pydantic==2.11.7
enum34==1.1.10
dagshub==0.6.3
mlflow==3.3.2
feature_engine==1.9.3
xgboost==3.0.4
fastapi==0.116.1
redis==6.4.0
prometheus-fastapi-instrumentator==7.1.0